wxGUI/datacatalog: new checking functions in check.py created (moved from guiutils.py), also existing functions as get_reason_mapset_removable moved to checks.py. get_reason_grassdb_removable was not moved because of GetListOfLocations function. MapsetDialog and LocationDialog a bit changed, now it uses GenericValidator instread of MultiGenericValidator.
